revise test functions
