+++ Server/.gitignore
@@ -2,4 +2,3 @@ venv/
 env/
 *.iml
 *.xml
-
