--- Server/Image_module/.cache/v/pep8/mtimes
@@ -1,66 +1,0 @@
-{
-  "/Users/zhiwei/Final_project/ImageProcessorS18/Server/Image_module/__init__.py": [
-    1524607092.141107,
-    []
-  ],
-  "/Users/zhiwei/Final_project/ImageProcessorS18/Server/Image_module/contrast_stretching.py": [
-    1524852456.7745416,
-    []
-  ],
-  "/Users/zhiwei/Final_project/ImageProcessorS18/Server/Image_module/decode_image.py": [
-    1524801022.6058025,
-    []
-  ],
-  "/Users/zhiwei/Final_project/ImageProcessorS18/Server/Image_module/encode_image.py": [
-    1524765942.0620792,
-    []
-  ],
-  "/Users/zhiwei/Final_project/ImageProcessorS18/Server/Image_module/function_test/test_contrast_stretching.py": [
-    1524800904.1121457,
-    []
-  ],
-  "/Users/zhiwei/Final_project/ImageProcessorS18/Server/Image_module/function_test/test_decode_image.py": [
-    1524797630.3523126,
-    []
-  ],
-  "/Users/zhiwei/Final_project/ImageProcessorS18/Server/Image_module/function_test/test_histogram.py": [
-    1524800271.199659,
-    []
-  ],
-  "/Users/zhiwei/Final_project/ImageProcessorS18/Server/Image_module/function_test/test_histogram_equalization.py": [
-    1524850351.563914,
-    []
-  ],
-  "/Users/zhiwei/Final_project/ImageProcessorS18/Server/Image_module/function_test/test_log_compression.py": [
-    1524800904.1041172,
-    []
-  ],
-  "/Users/zhiwei/Final_project/ImageProcessorS18/Server/Image_module/function_test/test_reverse_video.py": [
-    1524800904.1095166,
-    []
-  ],
-  "/Users/zhiwei/Final_project/ImageProcessorS18/Server/Image_module/function_test/test_strip_image.py": [
-    1524799802.3686957,
-    []
-  ],
-  "/Users/zhiwei/Final_project/ImageProcessorS18/Server/Image_module/histogram_equalization.py": [
-    1524801022.6058297,
-    []
-  ],
-  "/Users/zhiwei/Final_project/ImageProcessorS18/Server/Image_module/image_histogram.py": [
-    1524838689.5695977,
-    []
-  ],
-  "/Users/zhiwei/Final_project/ImageProcessorS18/Server/Image_module/log_compression.py": [
-    1524852476.2561593,
-    []
-  ],
-  "/Users/zhiwei/Final_project/ImageProcessorS18/Server/Image_module/reverse_video.py": [
-    1524840099.67451,
-    []
-  ],
-  "/Users/zhiwei/Final_project/ImageProcessorS18/Server/Image_module/strip_image.py": [
-    1524795079.3586988,
-    []
-  ]
-}
