+++ Server/image_functions/image_module/encode_image.py
