+++ Server/image_functions/image_module/strip_image.py
